add datasets to project, remove duplicates and 0 and negative values

+++ bruteforce_utils.py
@@ -17,7 +17,7 @@ def load_data(name_of_file):
 
     # remove first list with header
     data_shares = data_shares[1:]
-    data_shares = [(a, int(b), float(c)) for a, b, c in data_shares]
+    data_shares = [(a, float(b), float(c)) for a, b, c in data_shares]
 
     return data_shares
